New server paradigm
Server's bipartisan nature is now reflected in the way URLs and files are structured. There are 2 server modes: (1) API and (2) file server. So, now there are 2 apps placed in packages 'api' and 'pub'. The 'api' package is responsible for the API calls while the 'pub' package simply serves files from the 'pub' folder and all of its subfolders.

Typical API call: /api/song/49
Typical pub call: /pub/lisn/index.html

In the 'pub' folder, there's also the 'fail' folder that contains static HTML files that are sent over to the client in an even of failure.